- mod_rewrite of avatars - changed mod_rewrite 'img' to 'card' - changed mod_rewrite 'stats' to 'info'
Also, caught an issue.
Add if the domain is a test server for the character_avatar.php page, since it tries to find the local images on the main server. If it's a test server, give a placeholder avatar to be able to test placing it appropriately.
Might want to add caching as well if you haven't already to prevent it being generated every time an avatar is requested.

Keep up the great work.

git-svn-id: https://subversion.assembla.com/svn/mapler-me@59 7cbc8f23-9b70-4854-b091-be65f742de0c