So.. I'm extremely confused.
The notice.txt is somehow saved in /inc/ and accessible through there, while the real one in /manage/ is blank, which was causing the failed check.

Although.. when attempting to include it, it showed as not existent (php error)..

if this doesn't fix it, not sure what other route to go.

git-svn-id: https://subversion.assembla.com/svn/mapler-me@459 7cbc8f23-9b70-4854-b091-be65f742de0c