Hey Erwin, I haven't been able to get on Whatsapp recently because I had my phone charger stolen and haven't been able to charge it since.
Hope you enjoy this revision as much as I did working on it.

- Added parsing for images hosted on our new CDN by using !idhere in statuses.
- All pictures will display beautifully on the right as a thumbnail and are clickable.

===Mapler.me CDN===
I spent several hours working on a page that will allow users to upload pictures for them to include in statuses (or just use in general). The page uses CURL to push the file as well as information to a remote page (http://cdn.mapler.me/add.php) which in return uploads them to cdn.mapler.me/media/.

For the time being the uploader has not been publicized as I have not added security (checking for only images) which I will do tomorrow or soon.

Filenames are also decided by creating an MD5 hash for the file's temporary name when being sent over PHP. When you can find / work on that increasing id script similar to bit.ly, it will replace this so there aren't long ugly ids.

When a user adds a new picture, it is logged under cdn_log with their account_id as well as the file's name.

Besides that, glad with how it came out, I'll clean it up a bit tomorrow and email me if there's anything important you've had to tell me or send me a DM as always. :)

git-svn-id: https://subversion.assembla.com/svn/mapler-me@604 7cbc8f23-9b70-4854-b091-be65f742de0c